requirements.txt
arrow==0.15.8
asgiref==3.2.10
beautifulsoup4==4.9.1
binaryornot==0.4.4
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
defusedxml==0.6.0
Django==3.1
django-allauth==0.42.0
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
lxml==4.5.1
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
